code segregation successful
