--- DBConnection.java
@@ -1,25 +1,0 @@
-import java.io.FileInputStream;
-import java.sql.*;
-import java.util.Properties;
-
-public class DBConnection {
-
-    public static Connection getConnection() {
-        try {
-            Properties props = new Properties();
-            FileInputStream fis = new FileInputStream("db.properties");
-            props.load(fis);
-
-            String url = props.getProperty("db.url");
-            String user = props.getProperty("db.user");
-            String password = props.getProperty("db.password");
-
-            Class.forName("com.mysql.cj.jdbc.Driver");
-            return DriverManager.getConnection(url, user, password);
-
-        } catch (Exception e) {
-            System.out.println("Database connection error: " + e);
-            return null;
-        }
-    }
-}
