+++ src/Balance.java
@@ -1,0 +1,5 @@
+public class Balance {
+    public static void checkBalance(double balance) {
+        System.out.println("Balance: " + balance);
+    }
+    }
